get rid of audioflamingo3 dependency
